requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
httpx
pydantic>=2
pydantic-settings
rich
pytest
